+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `compute_dwelltimes_using_poly` to `sarkit.cphd`
+
 ### Removed
 - Unused `_processing` module